fix start_here.sh to show more helpful errors

+++ tooling/rasberry_pi_setup_scripts/install_pkgs.sh
@@ -33,11 +33,13 @@ exe() { echo -e "$Black$> $@ $Color_Off" >&2; eval "$@" ; }
 
 # setup the rov-web folder to be a git repo inline with github
 exe "cd '$HOME/rov-web/'"
-exe "git init -b main"
-exe "git remote add origin 'https://github.com/KW-M/rov-web.git'"
-exe "git fetch"
-exe "git reset --hard origin/main"
-exe "git branch --set-upstream-to=origin/main main"
+if [ ! -e ".git/" ]; then
+    exe "git init -b main"
+    exe "git remote add origin 'https://github.com/KW-M/rov-web.git'"
+    exe "git fetch"
+    exe "git reset --hard origin/main"
+    exe "git branch --set-upstream-to=origin/main main"
+fi
 
 # --------- Update System Packages ------------
 # From: https://learn.adafruit.com/circuitpython-on-raspberrypi-linux/installing-circuitpython-on-raspberry-pi
